+++ src/test/java/org/opensearch/flowframework/workflow/NoOpStepTests.java
@@ -15,6 +15,9 @@ import org.opensearch.test.OpenSearchTestCase;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.Map;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicReference;
 
 import static org.opensearch.flowframework.common.CommonValue.DELAY_FIELD;
 
@@ -48,6 +51,49 @@ public class NoOpStepTests extends OpenSearchTestCase {
         assertTrue(future.isDone());
         // Sleep isn't exactly accurate so leave 100ms of roundoff
         assertTrue(System.nanoTime() - start > 900_000_000L);
+    }
+
+    public void testNoOpStepInterrupt() throws IOException, InterruptedException {
+        NoOpStep noopStep = new NoOpStep();
+        WorkflowData delayData = new WorkflowData(Map.of(DELAY_FIELD, "5s"), null, null);
+
+        CountDownLatch latch = new CountDownLatch(1);
+        // Fetch errors from the separate thread
+        AtomicReference<AssertionError> assertionError = new AtomicReference<>();
+
+        Thread testThread = new Thread(() -> {
+            try {
+                PlainActionFuture<WorkflowData> future = noopStep.execute(
+                    "nodeId",
+                    delayData,
+                    Collections.emptyMap(),
+                    Collections.emptyMap(),
+                    Collections.emptyMap()
+                );
+                try {
+                    future.actionGet();
+                } catch (Exception e) {
+                    // Ignore the IllegalStateExcption/InterruptedExcpetion
+                }
+                assertTrue(future.isDone());
+                assertTrue(future.isCancelled());
+                assertTrue(Thread.currentThread().isInterrupted());
+            } catch (AssertionError e) {
+                assertionError.set(e);
+            } finally {
+                latch.countDown();
+            }
+        });
+
+        testThread.start();
+        Thread.sleep(100);
+        testThread.interrupt();
+
+        latch.await(1, TimeUnit.SECONDS);
+
+        if (assertionError.get() != null) {
+            throw assertionError.get();
+        }
     }
 
     public void testNoOpStepParse() throws IOException {